+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/data/storage/GalleryPreferences.kt
@@ -5,4 +5,5 @@ import ua.com.radiokot.photoprism.features.gallery.data.model.GalleryItemScale
 
 interface GalleryPreferences {
     val itemScale: BehaviorSubject<GalleryItemScale>
+    var livePhotosAsImages: BehaviorSubject<Boolean>
 }