+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/data/storage/GalleryPreferencesOnPrefs.kt
@@ -5,6 +5,7 @@ import io.reactivex.rxjava3.subjects.BehaviorSubject
 import ua.com.radiokot.photoprism.extension.kLogger
 import ua.com.radiokot.photoprism.extension.tryOrNull
 import ua.com.radiokot.photoprism.features.gallery.data.model.GalleryItemScale
+import ua.com.radiokot.photoprism.util.booleanPreferenceSubject
 import ua.com.radiokot.photoprism.util.stringifyPreferenceSubject
 
 class GalleryPreferencesOnPrefs(
@@ -31,5 +32,18 @@ class GalleryPreferencesOnPrefs(
                     }
                 }
             },
+        )
+
+    override var livePhotosAsImages: BehaviorSubject<Boolean> =
+        booleanPreferenceSubject(
+            preferences = preferences,
+            key = "${keyPrefix}_lp_as_images",
+            defaultValue = false,
+            onValuePut = { _, newValue ->
+                log.debug {
+                    "livePhotosAsImages::onNext(): set_value:" +
+                            "\nvalue=$newValue"
+                }
+            }
         )
 }